Store aboard/fatalities info separately
The information about aboard/fatalities was stored in a string
which was in a format "_total_ (passengers:_num_, crew=_num_)".

This made it hard to get specific information about the aboard
and fatalties data e.g. passengers aboard in a specific plane.

I've modified the scraping function to parse the strings and treat
each value as a separate column i.e. there are now 6 columns. They
are: all_aboard, passengers_aboard, crew_aboard, all_fatalities,
passenger_fatalities and crew_fatalities.

This should make it easy for people analyzing the data.